Refactor SVC handler

+++ Ryujinx.HLE/HOS/Kernel/KernelTransfer.cs
@@ -4,14 +4,14 @@ namespace Ryujinx.HLE.HOS.Kernel
 {
     static class KernelTransfer
     {
-        public static bool UserToKernelInt32(Horizon system, long address, out int value)
+        public static bool UserToKernelInt32(Horizon system, ulong address, out int value)
         {
             KProcess currentProcess = system.Scheduler.GetCurrentProcess();
 
-            if (currentProcess.CpuMemory.IsMapped(address) &&
-                currentProcess.CpuMemory.IsMapped(address + 3))
+            if (currentProcess.CpuMemory.IsMapped((long)address) &&
+                currentProcess.CpuMemory.IsMapped((long)address + 3))
             {
-                value = currentProcess.CpuMemory.ReadInt32(address);
+                value = currentProcess.CpuMemory.ReadInt32((long)address);
 
                 return true;
             }
@@ -21,14 +21,14 @@ namespace Ryujinx.HLE.HOS.Kernel
             return false;
         }
 
-        public static bool UserToKernelString(Horizon system, long address, int size, out string value)
+        public static bool UserToKernelString(Horizon system, ulong address, int size, out string value)
         {
             KProcess currentProcess = system.Scheduler.GetCurrentProcess();
 
-            if (currentProcess.CpuMemory.IsMapped(address) &&
-                currentProcess.CpuMemory.IsMapped(address + size - 1))
+            if (currentProcess.CpuMemory.IsMapped((long)address) &&
+                currentProcess.CpuMemory.IsMapped((long)address + size - 1))
             {
-                value = MemoryHelper.ReadAsciiString(currentProcess.CpuMemory, address, size);
+                value = MemoryHelper.ReadAsciiString(currentProcess.CpuMemory, (long)address, size);
 
                 return true;
             }
@@ -38,14 +38,14 @@ namespace Ryujinx.HLE.HOS.Kernel
             return false;
         }
 
-        public static bool KernelToUserInt32(Horizon system, long address, int value)
+        public static bool KernelToUserInt32(Horizon system, ulong address, int value)
         {
             KProcess currentProcess = system.Scheduler.GetCurrentProcess();
 
-            if (currentProcess.CpuMemory.IsMapped(address) &&
-                currentProcess.CpuMemory.IsMapped(address + 3))
+            if (currentProcess.CpuMemory.IsMapped((long)address) &&
+                currentProcess.CpuMemory.IsMapped((long)address + 3))
             {
-                currentProcess.CpuMemory.WriteInt32ToSharedAddr(address, value);
+                currentProcess.CpuMemory.WriteInt32ToSharedAddr((long)address, value);
 
                 return true;
             }
@@ -53,14 +53,14 @@ namespace Ryujinx.HLE.HOS.Kernel
             return false;
         }
 
-        public static bool KernelToUserInt64(Horizon system, long address, long value)
+        public static bool KernelToUserInt64(Horizon system, ulong address, long value)
         {
             KProcess currentProcess = system.Scheduler.GetCurrentProcess();
 
-            if (currentProcess.CpuMemory.IsMapped(address) &&
-                currentProcess.CpuMemory.IsMapped(address + 7))
+            if (currentProcess.CpuMemory.IsMapped((long)address) &&
+                currentProcess.CpuMemory.IsMapped((long)address + 7))
             {
-                currentProcess.CpuMemory.WriteInt64(address, value);
+                currentProcess.CpuMemory.WriteInt64((long)address, value);
 
                 return true;
             }
